Map Rotation
Unfortunatly I couldn't find a way to only rotate the character. There were way too many problems that I ran into with multiple objects trying to rotate the character (camera problems).

So instead I decided to go with rotating the whole map, This is bad in the sense that it may lag a lot during the rotation. Hopefully with how low poly the map should be, it shouldn't be a problem.